Fix `connect!`
`connect!` did the opposite of what was described in the docstring.
The docstring is more consistent with the convention of the modified
argument being first, so we go with the docstring and change the code.

While technically this is a bugfix, it could trigger StackOverFlow
errors, so let's call this a breaking change.