Adjust specialization of `on` and `map`
It's not entirely obvious whether this is the right thing to do, but the
idea here is that by ensuring that `on` and `map` get specialized
for their function argument, it makes everything more precompilable.

It might not be as important with precompilation of observables, so it's
worth being a bit skeptical about this.