Rework `map` and `map!`
There are several things that are a bit complicated/confusing about
`map` and `map!`:
- the `init` keyword also affects whether `f` runs (controlling the initial
  value is not orthogonal to execution, which spells trouble for
  functions with side effects)
- in contrast with `Base.map!`, `map!(f, dest::Observable, args..)`
  does not eagerly update `dest`---that doesn't happen until one of
  `args` updates
- `map` always yields an `Observable{Any}`, somewhat in contradiction
  to the behavior of `Base.map` (see #34)

This redesigns `map` and `map!` to address these issues (fixes #34).
It is a breaking change, although I've put in a depwarn for the case
where `init` was supplied. (The non-deprecated changes are so fundamental
that this may not be useful, however.)

Now:
- `map` always runs `f` on initial creation
- `map!` typically runs `f` on the initial call, but it can be
  suppressed with an `update=false` keyword
- `map` creates an `Observable{T}` with specific `T`, but users can
  control output `eltype` by using `map!(f, Observable{S}(), args...)`
  instead.

Using the new API, the old `map` behavior can be obtained with
    map!(f, Observable{Any}(), args...)
or with an `init` it's
    map!(f, Observable{Any}(init), args...; update=false)