Adjust specialization of `on` and `map` (#61)
The idea here is to preserve specialization in cases where it will be used,
and drop specialization when it won't. The creation of `OnUpdate` and `MapUpdater`
require specializing on the type, so it makes sense to force specialization up
until the point where these objects are created. Conversely, `on` and `off`
can't make use of specialized information about `f`, so we drop it as quickly
as we can to avoid the cost of inference and codegen.